Implement `Clone` for `HrirSphere`
Thanks again for publishing this as a separate crate! I've been having
fun doing [some
experiments](https://github.com/nannou-org/nannou/pull/669) in an
attempt to better understand the crate.

I'm now attempting to apply the crate to a downstream project. The
project involves a 5-channel surround sound installation, and I'm using
`hrtf` to allow me to "fake" the surround effect while I'm mixing on my
headphones for times when I don't have access to the speakers. To do
this, my plan is to create a `HrtfProcessor` per speaker, where each
will have a static position at evenly distributed angles around the
listener.

I noticed that the `HrtfProcessor` requires ownership over its
`HrirSphere`. Currently, this means re-reading the same `HrirSphere`
5-times, even though the data is the same. With this commit, it is
possible to load the sphere once and clone it as required for each
processor.

*I hope the whitespace trimming in the diff is OK - my editor is setup
to do this automatically on save!*